server/requirements.txt
flask==2.3.3
flask-cors==4.0.0
PyPDF2==3.0.1
python-dotenv==1.0.0
openai==1.3.9
requests==2.31.0
httpx==0.25.2
